Packages, Menu and Win Condition
Installed the post processing package and the textmeshpro package. Setup both the main menu and a working pause menu. There is now a win condition and a lose condition, where if either the starfish or the coral population reaches 0, the game ends and gives a message based on which population reached 0 first